% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpack3.R
\name{cheat}
\alias{cheat}
\title{cheat_function allows users to cheat on assignment 3.1}
\usage{
cheat(number)
}
\arguments{
\item{number}{numeric variable}
}
\description{
cheat_function allows users to cheat on assignment 3.1
}
